DDLLogEntry.parse() to populate query and initiator
